QUnit tests: get rid of material testing where possible (#31459)
