--- packages/devextreme/testing/tests/DevExpress.ui.widgets.pivotGrid/pivotGrid.genericTheme.tests.js
@@ -1,4 +1,0 @@
-import 'generic_light.css!';
-
-import { runThemesSharedTests } from './pivotGrid.themes.sharedTests.js';
-runThemesSharedTests('generic_light');
