--- packages/devextreme/testing/tests/DevExpress.ui.widgets/accordion.genericTheme.tests.js
@@ -1,4 +1,0 @@
-import 'generic_light.css!';
-
-import { runThemesSharedTests } from './accordionParts/accordion.themes.sharedTests.js';
-runThemesSharedTests('generic_light');
