--- packages/devextreme/testing/tests/DevExpress.ui.widgets/toolbar.genericTheme.tests.js
@@ -1,4 +1,0 @@
-import 'generic_light.css!';
-
-import { runThemesSharedTests } from './toolbarParts/toolbar.themes.sharedTests.js';
-runThemesSharedTests('generic_light');
