--- packages/devextreme/testing/tests/DevExpress.ui.widgets/toolbar.materialTheme.tests.js
@@ -1,4 +1,0 @@
-import 'material_blue_light.css!';
-
-import { runThemesSharedTests } from './toolbarParts/toolbar.themes.sharedTests.js';
-runThemesSharedTests('material_blue_light');
